fix: move bitToBooleanArray to rpc/utils
