+++ packages/cli/src/rpc/modules/portal.ts
@@ -20,7 +20,7 @@ import {
 
 import { BEACON_CLIENT_NOT_INITIALIZED, CONTENT_NOT_FOUND, INVALID_PARAMS } from '../error-code.js'
 import { content_params } from '../schema/index.js'
-import { callWithStackTrace, isValidId } from '../util.js'
+import { bitToBooleanArray, callWithStackTrace, isValidId } from '../util.js'
 import { middleware, validators } from '../validators.js'
 
 import { BitArray } from '@chainsafe/ssz'
@@ -29,35 +29,7 @@ import type { Debugger } from 'debug'
 import type { BeaconNetwork, HistoryNetwork, PortalNetwork, StateNetwork } from 'portalnetwork'
 import type { GetEnrResult } from '../schema/types.js'
 
-/**
- * Converts a BitArray response to a boolean array indicating which content keys were accepted
- * @param res The response from sendOffer (can be BitArray, undefined, or empty array)
- * @param contentKeysLength The number of content keys that were offered
- * @returns Object with success array, declined flag, or failed flag
- */
-function bitToBooleanArray(res: any, contentKeysLength: number): { success?: boolean[]; declined?: boolean; failed?: boolean } {
-  if (res === undefined) {
-    return { declined: true }
-  }
-  
-  if (Array.isArray(res) && res.length === 0) {
-    return { declined: true }
-  }
-  
-  // If res is a BitArray, convert it to boolean array
-  if (res !== undefined && res !== null && typeof res === 'object' && 'getTrueBitIndexes' in res) {
-    const acceptedBits = (res as any).getTrueBitIndexes()
-    const successArray = new Array(contentKeysLength).fill(false)
-    acceptedBits.forEach((index: number) => {
-      if (index < contentKeysLength) {
-        successArray[index] = true
-      }
-    })
-    return { success: successArray }
-  }
-  
-  return { success: new Array(contentKeysLength).fill(true) }
-}
+
 
 const methods = [
   // state
